sms-gateway/requirements.txt
httpx>=0.24.0
python-dotenv>=1.0.0
loguru>=0.5.3
fastapi>=0.68.0
uvicorn>=0.15.0 